fix: revise earlystopping callback's min_delta parameter

+++ src/TensorFlowNET.Keras/Callbacks/Earlystopping.cs
@@ -11,7 +11,7 @@ namespace Tensorflow.Keras.Callbacks;
 public class EarlyStopping: ICallback
 {
     int _paitence;
-    int _min_delta;
+    float _min_delta;
     int _verbose;
     int _stopped_epoch;
     int _wait;
@@ -26,7 +26,7 @@ public class EarlyStopping: ICallback
     CallbackParams _parameters;
     public Dictionary<string, List<float>>? history { get; set; }
     // user need to pass a CallbackParams to EarlyStopping, CallbackParams at least need the model
-    public EarlyStopping(CallbackParams parameters,string monitor = "val_loss", int min_delta = 0, int patience = 0,
+    public EarlyStopping(CallbackParams parameters,string monitor = "val_loss", float min_delta = 0f, int patience = 0,
         int verbose = 1, string mode = "auto", float baseline = 0f, bool restore_best_weights = false,
         int start_from_epoch = 0)
     {
